requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
aiosqlite==0.21.0
alembic==1.17.0
annotated-doc==0.0.2
annotated-types==0.7.0
anyio==4.11.0
asgiproxy==0.2.0
attrs==25.4.0
bcrypt==5.0.0
bidict==0.23.1
boto3==1.40.58
botocore==1.40.58
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cryptography==46.0.3
databricks-sdk==0.70.0
databricks-sqlalchemy==2.0.1
distro==1.9.0
fastapi==0.120.0
frozenlist==1.8.0
git-remote-s3==0.3.0
google-auth==2.41.1
granian==2.5.5
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.11.1
jmespath==1.0.1
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
multidict==6.7.0
narwhals==2.9.0
numpy==2.3.4
openai==2.6.0
packaging==25.0
pandas==2.3.3
pgvector==0.4.1
platformdirs==4.5.0
plotly==6.3.1
propcache==0.4.1
protobuf==6.33.0
psutil==7.1.1
psycopg==3.2.11
psycopg-binary==3.2.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
PyGithub
Pygments==2.19.2
PyJWT==2.10.1
PyNaCl==1.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.3
python-multipart==0.0.20
python-socketio==5.14.2
pytz==2025.2
redis==6.4.0
reflex==0.8.17a1
reflex-ag-grid==0.0.11
reflex-enterprise==0.3.5
reflex-global-hotkey==1.2.3
reflex-hosting-cli==0.1.58
requests==2.32.5
rich==14.2.0
rsa==4.9.1
s3transfer==0.14.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
sqlmodel==0.0.27
starlette==0.48.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchfiles==1.1.1
websockets==15.0.1
wrapt==2.0.0
wsproto==1.2.0
yarl==1.22.0